requirements.txt
flask
gunicorn
PyJWT
marshmallow
marshmallow_dataclass
apiflask[yaml]
Werkzeug

peewee
peewee-db-evolve
pymysql
psycopg2

python-dotenv
python-dateutil

beautifulsoup4
requests
urllib3
rq

redis
